test(openai): restore tool-call regression to HTTP test suite
Move the non-streaming tool-call ordering regression back into
TestOpenAIHTTPBackend and reuse its existing request-handler fixture.
Remove the standalone test module and its duplicated setup.

Preserve both error_stop and ignore_stop cases, all upstream HTTP tests,
and the existing runtime fix without behavioral changes.

Generated-by: ChatGPT
