+++ tests/unit/backends/openai/test_http.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 import asyncio
 import json
 from contextlib import nullcontext
+from typing import Literal
 from unittest.mock import MagicMock, Mock, patch
 
 import httpx
@@ -844,6 +845,70 @@ class TestOpenAIHTTPBackend:
         assert "max_tokens" not in sent_body  # None value filtered
         assert "top_p" not in sent_body  # None value filtered
         assert "stream" not in sent_body  # None value filtered
+
+    @pytest.mark.regression
+    @pytest.mark.asyncio
+    @async_timeout(10.0)
+    @pytest.mark.parametrize(
+        ("behavior", "expected_error", "expected_yields"),
+        [
+            ("error_stop", ValueError, 0),
+            ("ignore_stop", asyncio.CancelledError, 1),
+        ],
+    )
+    async def test_non_streaming_missing_tool_call_stop_behavior(
+        self,
+        httpx_mock: HTTPXMock,
+        mock_request_handler,
+        behavior: Literal["ignore_stop", "error_stop"],
+        expected_error: type[BaseException],
+        expected_yields: int,
+    ):
+        """Missing tool calls match streaming stop behavior before termination.
+
+        ## WRITTEN BY AI ##
+        """
+        httpx_mock.add_response(
+            url="http://test/v1/chat/completions",
+            json={"choices": [{"message": {"content": "no tool call"}}]},
+        )
+        backend = _make_backend(
+            target="http://test",
+            model="test-model",
+            stream=False,
+            validate_backend=False,
+            request_format="/v1/chat/completions",
+            tool_call_missing_behavior=behavior,
+        )
+        await backend.process_startup()
+        request = GenerationRequest(
+            columns={"text_column": ["call the tool"]},
+            turn_type="client_tool_call",
+        )
+        request_info = RequestInfo(
+            request_id="test-id",
+            status="pending",
+            scheduler_node_id=1,
+            scheduler_process_id=1,
+            scheduler_start_time=123.0,
+            timings=RequestTimings(),
+        )
+        mock_handler, handler_patch = mock_request_handler
+        mock_handler.compile_non_streaming.return_value = GenerationResponse(
+            request_id="test-id",
+            request_args="test args",
+            text="no tool call",
+        )
+        yielded = []
+
+        async def consume_response():
+            async for item in backend.resolve(request, request_info):
+                yielded.append(item)
+
+        with handler_patch, pytest.raises(expected_error, match="tool call"):
+            await consume_response()
+
+        assert len(yielded) == expected_yields
 
 
 class TestOpenAIBackendToolCallMissingBehavior:
